+++ internal/cmd/commands/hostcatalogs/funcs.go
@@ -1,6 +1,8 @@
 package hostcatalogs
 
 import (
+	"fmt"
+	"sort"
 	"time"
 
 	"github.com/hashicorp/boundary/api/hostcatalogs"
@@ -32,6 +34,31 @@ func generateHostCatalogTableOutput(in *hostcatalogs.HostCatalog) string {
 		"",
 		"  Scope:",
 		base.ScopeInfoForOutput(in.Scope, maxLength),
+	}
+
+	if len(in.AuthorizedActions) > 0 {
+		ret = append(ret,
+			"",
+			"  Authorized Actions:",
+			base.WrapSlice(4, in.AuthorizedActions),
+		)
+	}
+
+	if len(in.AuthorizedCollectionActions) > 0 {
+		keys := make([]string, 0, len(in.AuthorizedCollectionActions))
+		for k := range in.AuthorizedCollectionActions {
+			keys = append(keys, k)
+		}
+		sort.Strings(keys)
+		ret = append(ret, "",
+			"  Authorized Actions on Host Catalog's Collections:",
+		)
+		for _, key := range keys {
+			ret = append(ret,
+				fmt.Sprintf("    %ss:", key),
+				base.WrapSlice(6, in.AuthorizedCollectionActions[key]),
+			)
+		}
 	}
 
 	if len(in.Attributes) > 0 {